+++ src/features/wikis/tool-definitions.ts
@@ -5,6 +5,7 @@ import { GetWikiPageSchema } from './get-wiki-page/schema';
 import { CreateWikiSchema } from './create-wiki/schema';
 import { UpdateWikiPageSchema } from './update-wiki-page/schema';
 import { ListWikiPagesSchema } from './list-wiki-pages/schema';
+import { CreateWikiPageSchema } from './create-wiki-page/schema';
 
 /**
  * List of wikis tools
@@ -34,5 +35,11 @@ export const wikisTools: ToolDefinition[] = [
     name: 'list_wiki_pages',
     description: 'List pages within an Azure DevOps wiki',
     inputSchema: zodToJsonSchema(ListWikiPagesSchema),
+  },
+  {
+    name: 'create_wiki_page',
+    description:
+      'Create a new page in a wiki. If the page already exists at the specified path, it will be updated.',
+    inputSchema: zodToJsonSchema(CreateWikiPageSchema),
   },
 ];